create watchers file & move localStorageWatcher
